Update EffectReplantCrops.kt
This PR fixes an issue whereby if both `consume_seeds` and `only_fully_grown` are true, if a player breaks a non-fully-grown crop, it: breaks the crop, does not replant, drops the seed but ALSO consumes a seed. Meaning for each ungrown crop broken, the player loses a seed each time.