gave up with fully grown
The issue is that some value are Int, some are String and some are CropConfig.
These are not comparable, for example, the maxPoints returns an Int based on the config, eg: 4
The stages have an ID that is also a number, but the ID of them is a strong, "crop_stage_4". This is uncomarable. You;d think you could just split it by "_" and take the last character, but no, as this doesn't always work. Take Tomato by default, it has maxPoints of 6, but the point it gives tomato drops is stage 4, the others are for variations and oversized ones. This just makes it complicated and needs someone smarter than myself, ChatGPT and DeepSeek who understands this shit better than me.