fixed additional blank line
Currently, 'send_message' & 'broadcast' add a blank line message at the end when using plural key (messages). The 'run_command' and 'run_player_command' also run a blank command on every trigger, which in some cases produces a libreforge error.

This PR fixes all of the above and it is now much cleaner to use.