Added `villager_trade_multiplier`
This effect multiplies the cost to trade with a villager.
There are a few limitations, such as losing the GUI title (profession + level) whilst effect is active, once effect is off, it resumes as usual.

This is potentially still Experimental but worth adding anyway as functionally it seems to work